Add lcm() to reusable file.
Add unit tests for lcm() top-level function.

Refactor gcd() top-level function to use Long.

+++ src/main/kotlin/batch1/Problem9.kt
@@ -48,7 +48,7 @@ class SpecialPythagoreanTriplet {
                 }
                 var k = if (m % 2 == 1) m + 2 else m + 1
                 while (k < 2 * m && k <= kMax) {
-                    if (kMax % k == 0 && gcd(k, m) == 1) {
+                    if (kMax % k == 0 && gcd(k.toLong(), m.toLong()) == 1L) {
                         val triplet = pythagoreanTriplet(m, k - m, nMax / (k * m))
                         if (triplet.sum() == n) {
                             val product = triplet.product()
